Client data fetching and Static Data fetching
Implemented a page where you can search for Artists using Spoitfy's Web
API. I had implemented a utility module called spotifyAuth to retrieve
my app's access token, but I learned it was inaccessible from a Client
Component. This is because the utility modules are JavaScript that live
on the server and NextJS does not send it to the client alongisde the
component.

The workaround was making a NextJS api route to act as an interface for
fetching from the Spotify API.

I also implemented impelmented a Reggaeton track page that recommends
tracks using the API. This page shows how static data fetching with
revalidation in a Server Component works.

This endpoint in Spotify's API always gives you different results when
you call it. For this page component, I passed in the next:
{revalidation: 5} option into my fetch(). Refreshing the page won't show
you different results until 5 seconds pass. When the Server component
recieves new data from Spotify, NextJS rerenders the component.

IMPORTANT: This 5 second refresh of the Server component only happens in
a production build. When running npm run dev, the component will be
rerendered on every refresh.